[#15] refactor: HistoryManager 타입 생성 및 분리
