+++ ProjectManager/ProjectManager/Model/BoardManager.swift
@@ -1,22 +1,5 @@
 import Foundation
 
-enum HistoryLog {
-    case add(String)
-    case move(String, String, String)
-    case delete(String)
-    
-    var description: String {
-        switch self {
-        case .add(let title):
-            return "Added \(title)"
-        case .move(let title, let before,  let after):
-            return "Moved \(title) from \(before) to \(after)"
-        case .delete(let title):
-            return "Deleted \(title)"
-        }
-    }
-}
-
 class BoardManager {
     static let shared = BoardManager()
     
@@ -26,8 +9,6 @@ class BoardManager {
     lazy var boards: [Board] = {
         return [todoBoard, doingBoard, doneBoard]
     }()
-    
-    var historyContainer = [(String, Date)]()
     
     private var items: [Item] = []
