+++ ProjectManager/ProjectManager/ViewController/HistoryViewController.swift
@@ -5,7 +5,7 @@ class HistoryViewController: UIViewController {
 }
 extension HistoryViewController: UITableViewDataSource {
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return boardManager.historyContainer.count
+        return historyManager.historyContainer.count
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
